backend/requirements.txt
fastapi
uvicorn
pydantic
openai
httpx
python-dotenv
tinyfish
